Revert policy gateway docker compose to use image of policy token gateway
I don't like the idea of other components needing to build each other to test.  Much better to pull an image.  But I understand it was required in this case to test it because the changes in the policy gateway depeneded on the updated policy\_token\_gateway().

I'm not sure how best to handle this in the future.
Option 1: We typically use "latest" built images I'm thinking the PR should use the image.  Because the PR itself will trigger a build.  But this of course could cause issues if another PR is made from another branch.

Option 2: Don't update the compose file to build but putting a comment in the PR that it is required to test it.  Similar to Option 1 but has the reviewer/tester change the file themselves to test.

Option 3: As this was done but a tester reverts the change before accepting the PR.

For now I've changed it back to pull the image directly.